# docs/swagger.yaml
openapi: 3.0.3
info:
  title: MindFlow API
  version: 1.0.0
  description: Comprehensive API documentation for the Mental Health Trend Analytics Platform.
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation failed or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials or wrong role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/entry:
    get:
      tags: [Entries]
      summary: Get recent mood entries
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recent entries returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MoodEntry'
    post:
      tags: [Entries]
      summary: Create a mood entry for the logged-in user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoodEntryRequest'
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/entry/manual:
    post:
      tags: [Entries]
      summary: Create a mood entry for a patient as psychologist or admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/MoodEntryRequest'
                - type: object
                  required: [userId]
                  properties:
                    userId:
                      type: string
      responses:
        '201':
          description: Manual entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoodEntry'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden for patient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenResponse'
...
  /api/personal/notes:
    get:
      tags: [Notes]
      summary: Get private notes for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PersonalNote'
    post:
      tags: [Notes]
      summary: Create a private note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteRequest'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalNote'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/personal/notes/{id}:
    delete:
      tags: [Notes]
      summary: Delete a private note by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Note removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Note does not belong to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
    ForbiddenResponse:
      type: object
      properties:
        message:
          type: string
        alert:
          type: string
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
    Settings:
      type: object
      properties:
        notifications:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            weeklyReports:
              type: boolean
            patientAlerts:
              type: boolean
            systemUpdates:
              type: boolean
        appearance:
          type: object
          properties:
            theme:
              type: string
            colorScheme:
              type: string
            fontSize:
              type: string
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [patient, psychologist, admin]
    LoginRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [patient, psychologist, admin]
    AuthResponse:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        bio:
          type: string
        profileImage:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/Settings'
        role:
          type: string
        token:
          type: string
...
    MoodEntryRequest:
      type: object
      required: [emotion, category]
      properties:
        emotion:
          type: number
          minimum: 1
          maximum: 10
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
          enum: [General, Work, Social, Health, Family, Personal]
        notes:
          type: string
        timestamp:
          type: string
          format: date-time
    MoodEntry:
      type: object
      properties:
        _id:
          type: string
        user:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                username:
                  type: string
                email:
                  type: string
        emotion:
          type: number
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
        notes:
          type: string
        timestamp:
          type: string
          format: date-time
...
    NoteRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
    PersonalNote:
      type: object
      properties:
        _id:
          type: string
        user:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isDraft:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time